Day 16 solutions (Odin).
I think I significantly over-complicated this one.

The initial parsing of the data into a convenient graph structure
was very awkward. Once I had that done Part 1 was easy.

For Part 2 I switched to operating on the dual graph because I had
some accounting problems on vertices (does the turn cost happen before
or after reaching the vertex?).

Then I added a cache so for each edge we knew the optimal edges that
fed into it (can be more than one for multiple paths).

The backtracking went smoothly, but then I wasted a lot of time because
I had a bug when accounting for vertices that were visited multiple
times.